@@ -2,11 +2,11 @@
 
     <PropertyGroup>
         <OutputType>Exe</OutputType>
-        <TargetFramework>netcoreapp3.1</TargetFramework>
+        <TargetFramework>net5.0</TargetFramework>
     </PropertyGroup>
 
     <ItemGroup>
-        <ProjectReference Include="..\Collections\Collections.csproj"/>
+      <PackageReference Include="Microsoft.Data.SqlClient" Version="4.0.0-preview2.21264.2" />
     </ItemGroup>
 
 </Project>
