@@ -2,7 +2,7 @@ using System;
 
 namespace CovariantContrVariantDelegates
 {
-    class Program
+    internal class Program
     {
         public static void Local_Main()
         {
@@ -14,7 +14,7 @@ namespace CovariantContrVariantDelegates
             Console.WriteLine(message);
         }
 
-        private static void DisplayRedMessage(String message)
+        private static void DisplayRedMessage(string message)
         {
             // Устанавливаем красный цвет символов
             Console.ForegroundColor = ConsoleColor.Red;
@@ -24,29 +24,28 @@ namespace CovariantContrVariantDelegates
         }
     }
 
-    class CoVariant
+    internal class CoVariant
     {
-        class Person
+        private class Person
         {
             public string Name { get; set; }
         };
 
-        class Client : Person
+        private class Client : Person
         {
             public int money = 500;
         };
 
-        delegate Person PersonFactory(string name);
+        private delegate Person PersonFactory(string name);
 
         public static void Do()
         {
             PersonFactory personDel;
             personDel = BuildClient; // ковариантность
-            Person p = personDel("Tom"); //personDel("Tom") = Client
-            Console.WriteLine(p.Name); 
+            var p = personDel("Tom"); //personDel("Tom") = Client
+            Console.WriteLine(p.Name);
             // Client p = personDel("Tom"); error, cant Person to Client
             //Мы не можем здесь писать Client p потому что хоть он и возращает Client, Но изза того что делегат Person? 
-
         }
 
         private static Client BuildClient(string name)
