@@ -1,12 +1,10 @@
 using System;
-using System.Collections;
-using System.Collections.Generic;
 
 namespace playground
 {
-    class Programm
+    internal class Programm
     {
-        static void Main()
+        private static void Main()
         {
             Console.WriteLine("asadasda");
         }
