@@ -1,18 +1,16 @@
-﻿using System;
-
-namespace LINQSamples
+﻿namespace LINQSamples
 {
     class Program
     {
         static void Main()
         {
-            BaseLINQOperationsDemos.RunAllMethods();
-            SkipTakeSumDistinctDemos.RunAllMethods();
-            ContainingElementsFunctionsDemos.RunAllMethods();
-            ActionsWithSeveralCollections.RunAllMethods();
-            GroupByFuncs.RunAllMethods();
-            AggregateFunctions.RunAllMethods();
-            DefferedAndImmediateTypes.RunAllMethods();
+            // BaseLINQOperationsDemos.RunAllMethods();
+            // SkipTakeSumDistinctDemos.RunAllMethods();
+            // ContainingElementsFunctionsDemos.RunAllMethods();
+            // ActionsWithSeveralCollections.RunAllMethods();
+            // GroupByFuncs.RunAllMethods();
+            // AggregateFunctions.RunAllMethods();
+            // DefferedAndImmediateTypes.RunAllMethods();
         }
     }
 }
