@@ -1,7 +1,10 @@
 using System;
 using System.Collections.Concurrent;
+using System.IO;
 using System.Linq;
+using System.Threading;
 using System.Threading.Tasks;
+using Newtonsoft.Json.Linq;
 
 namespace TAP.Exercises
 {
@@ -21,19 +24,40 @@ namespace TAP.Exercises
             _myConcurrentDictionary = new ConcurrentDictionary<int, int>();
         }
 
-        public void FillTheDict()
+        private void Increment(int key)
         {
-            var rand = new Random();
-            Parallel.For(0, 20, _ =>
+            _myConcurrentDictionary.AddOrUpdate(key, 1, (k, v) => Interlocked.Increment(ref v));
+        }
+
+        public void FillTheDictRandNumbers(int threadsCount = 20000)
+        {
+            var rand = new ThreadLocal<Random>(() => new Random()).Value;
+            Parallel.For(0, threadsCount, _ =>
             {
                 var randKey = rand.Next(0, 10);
-                _myConcurrentDictionary.AddOrUpdate(randKey, 1, (k, v) => ++v);
+                // _myConcurrentDictionary.AddOrUpdate(randKey, 1, (k, v) => Interlocked.Increment(ref v));
+                Increment(randKey);
+            });
+        }
+
+        public void FillTheDictFromJsonFile(string path, int threadsCount = 512)
+        {
+            if (threadsCount > 512 || threadsCount < 1)
+            {
+                throw new ArgumentOutOfRangeException("threadsCount is not less than 1 or greater than 512.");
+            }
+
+            var json = JArray.Parse(File.ReadAllText(path));
+
+            json.Children().AsParallel().WithDegreeOfParallelism(threadsCount).ForAll(token =>
+            {
+                Increment(token.Value<int>());
             });
         }
 
         public void ShowSortedDict()
         {
-            foreach (var pair in _myConcurrentDictionary.OrderBy(pair => pair.Key))//no locking
+            foreach (var pair in _myConcurrentDictionary.OrderBy(pair => pair.Key)) //no locking
             {
                 Console.WriteLine("{0} - {1}", pair.Key, pair.Value);
             }
