@@ -1,21 +1,23 @@
 using System;
 using System.Collections.Generic;
+using System.IO;
+using System.Threading;
+using Newtonsoft.Json.Linq;
 
 namespace TAP.Exercises
 {
     public class MySelfFillingDictionaryWithoutLocking
     {
         private Dictionary<int, int> _myDict;
-
         public MySelfFillingDictionaryWithoutLocking()
         {
             _myDict = new Dictionary<int, int>();
         }
 
-        public void FillTheDict()
+        public void FillTheDictRandNumbers(int iterCount)
         {
-            var rand = new Random();
-            for (int i = 0; i < 20; i++)
+            var rand = new ThreadLocal<Random>(() => new Random()).Value;
+            for (int i = 0; i < iterCount; i++)
             {
                 var randKey = rand.Next(0, 10);
                 if (_myDict.ContainsKey(randKey))
@@ -25,6 +27,22 @@ namespace TAP.Exercises
                 else
                 {
                     _myDict.Add(randKey, 1);
+                }
+            }
+        }
+
+        public void FillTheDictFromFile(string path)
+        {
+            var json = JArray.Parse(File.ReadAllText(path));
+            foreach (var token in json)
+            {
+                if (_myDict.ContainsKey(token.Value<int>()))
+                {
+                    _myDict[token.Value<int>()]++;
+                }
+                else
+                {
+                    _myDict.Add(token.Value<int>(), 1);
                 }
             }
         }
